src/main/java/com/tosken/rest/service/UserService.java
package com.tosken.rest.service;

import com.tosken.rest.persistence.dao.UserRepository;
import com.tosken.rest.persistence.model.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Sort;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

@Service
@Transactional
public class UserService {

    private final UserRepository userRepository;

    @Autowired
    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public List<User> getAll() {
        return userRepository.findAll(Sort.by("lastname", "firstname"));
    }
}
